update code from PR review notes
